Refactor (#11)
Refactored code base and added sequential functionality as well as fixed some bugs

* Toplevel wrapper for aig partitioning (#9)

* Create wrapper verilog for aig partitioning.

* Update include list

* sequential support to optimize partitions

* fixing bug with constant in partitions

* Updated LSOracle to preserve IO naming

* Fixed cmake bug

* Fixed travis test bug

* Updated Lorina extension to fix BLIF reader bug for large networks

* Updated Mockturtle to add balancing functionality and fixed refactoring bug for AIGs

* Delete test.txt

* Delete test.txt

* Made modifications to mockturtle to be able to read and optimize sequential aiger files

* Fixed travis test bug

* Fixed end-to-end testing

* Fixed end-to-end testing

* Refactored code base

* Delete test_aig_script.hpp

* Delete test_mig_script.hpp

* Update mockturtle.hpp

* Delete deep_learn_model.json

* Added new AIG recipe that uses DSD resynthesis for refactoring

* Added sequential support to the BLIF reader

* Fixed testing bug

* Added sequential support to BLIF writer and added name preservation to AIG reader and Verilog writer

* Cleaned up repo and added improved AIG opt recipe for mixed synthesis

* Removed unneeded libraries

* Removed unneeded directory

* Automated oracle command further

* Removed unused library

* Removed unneeded files

* Fixed minor bug

* Fixed bug with feedthrough nets and improperly named IO

+++ lib/mockturtle/include/mockturtle/io/bench_reader.hpp
@@ -1,5 +1,5 @@
 /* mockturtle: C++ logic network library
- * Copyright (C) 2018  EPFL
+ * Copyright (C) 2018-2019  EPFL
  *
  * Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
@@ -55,11 +55,11 @@ namespace mockturtle
  * - `create_node`
  *
    \verbatim embed:rst
-  
+
    Example
-   
+
    .. code-block:: c++
-   
+
       klut_network klut;
       lorina::read_bench( "file.bench", bench_reader( klut ) );
    \endverbatim
@@ -107,7 +107,7 @@ public:
   {
     if ( type.size() > 2 && std::string_view( type ).substr( 0, 2 ) == "0x" && inputs.size() <= 6u )
     {
-      kitty::dynamic_truth_table tt( inputs.size() );
+      kitty::dynamic_truth_table tt( static_cast<int>( inputs.size() ) );
       kitty::create_from_hex_string( tt, type.substr( 2 ) );
 
       std::vector<signal<Ntk>> input_signals;
